feat:FA-69

+++ Ban-Hang/src/app/pages/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { IProduct } from 'src/app/interface/product';
 import { ProductService } from 'src/app/service/product.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-list',
@@ -17,8 +18,26 @@ export class ProductListComponent {
     )
   }
   removeId(_id: any) {
-    this.productService.deleteProduct(_id).subscribe(() => {
-      this.products = this.products.filter(product => product._id !== _id)
+    Swal.fire({
+      title: 'Delete?',
+      text: "Bạn chắc là muốn xóa chứ ?",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.productService.deleteProduct(_id).subscribe(() => {
+          this.products = this.products.filter(product => product._id !== _id)
+        })
+        Swal.fire(
+          'Deleted!',
+          'Sản phẩm đã được xóa',
+          'success'
+        )
+      }
     })
+
   }
 }
